requirements.txt
altair==4.2.0
attrs==22.2.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.1
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
Faker==16.1.0
favicon==0.7.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
htbuilder==0.6.1
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
markdownlit==0.0.5
MarkupSafe==2.1.1
matplotlib==3.6.2
more-itertools==9.0.0
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
protobuf==3.20.3
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.14.0
pymdown-extensions==9.9
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==13.0.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
st-annotated-text==3.0.0
streamlit==1.16.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.4
streamlit-faker==0.0.2
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
vega-datasets==0.9.0
watchdog==2.2.1
zipp==3.11.0
